+++ src/components/AppLayout/index.ts
@@ -1,2 +1,3 @@
-export { default } from "./AppLayout";
 export * from "./AppLayout";
+export { default } from "./AppLayout";
+export * from "./TopNav";
