+++ src/new-apps/components/AppListCard/index.ts
@@ -1,2 +1,3 @@
 export * from "./AppListCard";
 export { default } from "./AppListCard";
+export * from "./AppLogo";
